rename InputLabel to Label [TESTS FAILING]
