+++ src/components/atoms/Label/Label.md
@@ -3,7 +3,7 @@ Label for input Components. This component is meant to be used with an input.
 For a more complex component check `TextField`.
 
 ```jsx
-import { InputLabel } from '@zopauk/react-components';
+import { Label } from '@zopauk/react-components';
 
-<InputLabel htmlFor="username">Username</InputLabel>;
+<Label htmlFor="username">Username</Label>;
 ```
